make gardener/dashboard/identity/dns-controller repository and image configurable
The "versions" node is now more powerful and allows changing the git repository
as well as the image repository and tag (independent of the git repo).
It is still completely optional.

The structure is as follows:

versions:
  gardener:
    repo: <git repo url>
    tag: <tag to checkout - overwrites branch/commit>
    branch: <branch to checkout - overwrites commit>
    commit: <commit to checkout>
    apiserver:
      image_repo: <image repo url>
      image_tag: <image tag>
    controller_manager:
      image_repo: <see above>
      image_tag: <see above>
  dashboard:
    repo: <see above>
    tag: <see above>
    branch: <see above>
    commit: <see above>
    image_repo: <see above>
    image_tag: <see above>
  identity:
    repo: <see above>
    tag: <see above>
    branch: <see above>
    commit: <see above>
    image_repo: <see above>
    image_tag: <see above>
  dns_controller:
    image_tag: <see above>
    image_repo: <see above>

For each repo, only one of tag/branch/commit should be specified, otherwise
some of them will be ignored (see above).

When overwriting some of these values, keep in mind that this project is
tailored towards a specific version and might not work if that version is
changed.