add installation-handler script
There are some scenarios in which it is somewhat problematic that garden-setup persists the state only on the local hard drive. To work around that, the installation-handler script allows to dump the state and export folders into kubernetes secrets in the cluster.

When activated, before any sow command, the state and export folders will be restored from the secrets. After the sow command, they will be stored in secrets again.