move git and chart-checkout plugin calls to prepare step
Although the gen folder can be completely recreated by deploying a component again, deleting components was not always possible without it. Usually the reason for this is that the helm plugin needs to render the charts to delete them and since the plugins are run in reversed order at delete, the git plugin couldn't checkout the charts before that (and it also didn't check out anything during delete).
The git plugin in sow has changed to also perform a checkout on delete, but to also make that happen before the helm plugin, the calls to the git plugin had to be moved to the component.yaml file.