dependency-versions file
Because the test machinery needs to know the repositories and versions of the Gardener extensions,
the old GARDENER_VERSION file has been expanded to include these.

Although the file has a '.yaml' file extension, it currently has to be a JSON file due to a lack
of yaml-to-json converters in the current pipeline job image. Will maybe be changed later.